* coders/meta.c (ReadMETAImage): Fixed reading ICM color profile files.  Due to a typo ICM color profiles were being stored as IPTC profiles.  This restores proper operation of the -profile option.
* magick/image.c (DescribeImage): If the image is DirectClass,
then don't compute the number of unique colors unless verbose is
greater than one.  This change is made since computing the number
of unique colors may take hours for some images.  The handling of
the -verbose argument is changed so that it is cumulative.
Specifying -verbose multiple times increases the level of
verbosity.


* magick/nt_base.c (NTreaddir): Fix write beyond buffer length
reported in SourceForge issue #1182003.  Only impacts Windows.