SVG/MVG: Fix clipping of polygons to image boundaries.
Clipping of polygons in DrawImage() would sometime result
in a starting pixel location that was greater than the
ending pixel location, causing a subsequent call to
GetImagePixelsEx() to fail due a column count <= 0.  Modified
the clipping code to eliminate this condition, and also to
return early if the polygon lies completely outside the image
boundaries.  Also fixed variable declarations from a previous
commit that were causing problems for the C89 compiler.