SVG: Multiple modifications to fix ticket #322.
NOTE: Overview follows.  See the ChangeLog for more complete
details.

Ticket #322 exposed multiple bugs in render.c associated with
rendering zero length open subpaths according to the SVG spec.
These bugs resulted in missing features in the output, most
notably filled circles and/or squares that the spec says
should be drawn for zero length segments when round or square
line endcaps are enabled.

Modified code in render.c so that closed shapes are specifically
tagged as such, replacing the previous policy of detecting
closed shapes by comparing the first and last points to see if
they are identical (within MagickEpsilon). The old policy
prevented open subpaths with the same first and last point from
being rendered properly (per the SVG spec) when round or square
endcaps were enabled.

Fixed a bug in render.c, function GetPixelOpacity(), that
computes the distance between a line and a segment (i.e.,
polygon edge). For zero length segments this resulted in a
divide by zero and incorrect output.

Modified the duplicate point elimination code in render.c,
function ConvertPrimitiveToPath(), so that the first and last
points are always preserved. In addition to being part of the
fix for ticket #322, this modification also corrected a bug in
which if the last point of a closed shape was eliminated as a
duplicate, the shape would no longer be closed, and small
"nubs" would be rendered at the shape boundary where the shape
would have been closed.

Per the SVG spec, zero length arc subpaths are now treated like
a zero length "line to" command to the current point.