SVG: Allow empty class definitions; allow undefined classes.
Class definitions defined within a <style> block may now be
empty. For a reference such as 'class="classname"', the
"classname" is now allowed to be undefined. These relaxed
conditions are not specifically called out in the SVG spec
as being either acceptable or unacceptable, but other SVG
renders (e.g., Chrome) handle them this way.